# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIOSK API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/leave": {
      "post": {
        "tags": [
          "leave"
        ],
        "summary": "Create Leave",
        "description": "multipart/form-data (so the OD letter can travel with the form).",
        "operationId": "create_leave_api_v1_leave_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_leave_api_v1_leave_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/leave/mine": {
      "get": {
        "tags": [
          "leave"
        ],
        "summary": "My Requests",
        "operationId": "my_requests_api_v1_leave_mine_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/leave/queue": {
      "get": {
        "tags": [
          "leave"
        ],
        "summary": "Queue",
        "description": "Counsellor: own students' requests waiting for them. HOD: department requests waiting for the HOD.\nThe signed-in user decides, never a counsellorId sent by the browser.",
        "operationId": "queue_api_v1_leave_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/leave/history": {
      "get": {
        "tags": [
          "leave"
        ],
        "summary": "History",
        "description": "Decisions made: by this counsellor, or at the HOD stage for the department.",
        "operationId": "history_api_v1_leave_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ALL",
                "APPROVED",
                "REJECTED"
              ],
              "type": "string",
              "default": "ALL",
              "title": "Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave/{leave_id}": {
      "get": {
        "tags": [
          "leave"
        ],
        "summary": "Get Leave",
        "operationId": "get_leave_api_v1_leave__leave_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Leave Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave/{leave_id}/letter": {
      "get": {
        "tags": [
          "leave"
        ],
        "summary": "Get Letter",
        "description": "The uploaded OD letter. Open with fetch() + the login token, then show it from a blob URL.",
        "operationId": "get_letter_api_v1_leave__leave_id__letter_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Leave Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave/{leave_id}/decision": {
      "post": {
        "tags": [
          "leave"
        ],
        "summary": "Decide",
        "operationId": "decide_api_v1_leave__leave_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Leave Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leave/{leave_id}/reassign": {
      "post": {
        "tags": [
          "leave"
        ],
        "summary": "Reassign",
        "operationId": "reassign_api_v1_leave__leave_id__reassign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "leave_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Leave Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_leave_api_v1_leave_post": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/LeaveType"
          },
          "category": {
            "type": "string",
            "maxLength": 40,
            "title": "Category"
          },
          "fromDate": {
            "type": "string",
            "format": "date",
            "title": "Fromdate"
          },
          "toDate": {
            "type": "string",
            "format": "date",
            "title": "Todate"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "residentialAddress": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Residentialaddress"
          },
          "eventName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventname"
          },
          "organizer": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizer"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "letter": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Letter"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "category",
          "fromDate",
          "toDate"
        ],
        "title": "Body_create_leave_api_v1_leave_post"
      },
...
      "Decision": {
        "type": "string",
        "enum": [
          "APPROVE",
          "REJECT"
        ],
        "title": "Decision"
      },
      "DecisionIn": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DecisionIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaveType": {
        "type": "string",
        "enum": [
          "LEAVE",
          "OD"
        ],
        "title": "LeaveType"
      },
...
      "ReassignIn": {
        "properties": {
          "counsellorId": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Counsellorid"
          }
        },
        "type": "object",
        "required": [
          "counsellorId"
        ],
        "title": "ReassignIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}